Separate folder for the tree
